+++ test/router/ExchangeRouter.ts
@@ -8,7 +8,6 @@ import { getDepositKeys } from "../../utils/deposit";
 import { getWithdrawalKeys } from "../../utils/withdrawal";
 import { handleDeposit } from "../../utils/deposit";
 import { hashString } from "../../utils/hash";
-import { errorsContract } from "../../utils/error";
 import { OrderType, DecreasePositionSwapType, getOrderKeys } from "../../utils/order";
 import * as keys from "../../utils/keys";
 
@@ -24,7 +23,6 @@ describe("ExchangeRouter", () => {
     exchangeRouter,
     ethUsdMarket,
     ethUsdSpotOnlyMarket,
-    wnt,
     usdc;
   const executionFee = expandDecimals(1, 18);
 
@@ -41,7 +39,6 @@ describe("ExchangeRouter", () => {
       exchangeRouter,
       ethUsdMarket,
       ethUsdSpotOnlyMarket,
-      wnt,
       usdc,
     } = fixture.contracts);
   });
@@ -333,63 +330,5 @@ describe("ExchangeRouter", () => {
       tx,
       label: "exchangeRouter.createWithdrawal",
     });
-  });
-
-  it("simulateExecuteLatestDeposit", async () => {
-    await usdc.mint(user0.address, expandDecimals(50 * 1000, 6));
-    await usdc.connect(user0).approve(router.address, expandDecimals(50 * 1000, 6));
-
-    const currentTimestamp = (await ethers.provider.getBlock()).timestamp + 2;
-
-    await expect(
-      exchangeRouter.connect(user0).multicall(
-        [
-          exchangeRouter.interface.encodeFunctionData("sendWnt", [depositVault.address, expandDecimals(11, 18)]),
-          exchangeRouter.interface.encodeFunctionData("sendTokens", [
-            usdc.address,
-            depositVault.address,
-            expandDecimals(50 * 1000, 6),
-          ]),
-          exchangeRouter.interface.encodeFunctionData("createDeposit", [
-            {
-              addresses: {
-                receiver: user1.address,
-                callbackContract: user2.address,
-                uiFeeReceiver: user3.address,
-                market: ethUsdMarket.marketToken,
-                initialLongToken: ethUsdMarket.longToken,
-                initialShortToken: ethUsdMarket.shortToken,
-                longTokenSwapPath: [],
-                shortTokenSwapPath: [],
-              },
-              minMarketTokens: 100,
-              shouldUnwrapNativeToken: true,
-              executionFee,
-              callbackGasLimit: "200000",
-              srcChainId: 0,
-              dataList: [],
-            },
-          ]),
-          exchangeRouter.interface.encodeFunctionData("simulateExecuteLatestDeposit", [
-            {
-              primaryTokens: [wnt.address, usdc.address],
-              primaryPrices: [
-                {
-                  min: expandDecimals(5000, 12),
-                  max: expandDecimals(5000, 12),
-                },
-                {
-                  min: expandDecimals(1, 24),
-                  max: expandDecimals(1, 24),
-                },
-              ],
-              minTimestamp: currentTimestamp,
-              maxTimestamp: currentTimestamp,
-            },
-          ]),
-        ],
-        { value: expandDecimals(11, 18) }
-      )
-    ).to.be.revertedWithCustomError(errorsContract, "EndOfOracleSimulation");
   });
 });
